requirements.txt
fire==0.7.0
numpy==2.2.0
opencv_python==4.10.0.84
Pillow==11.0.0
requests==2.32.3
scikit-image==0.24.0
torch==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.67.1
